Rename hyp3_daemon -> daemon in hyp3_process module
