+++ modules/hyp3_process/src/hyp3_process/__init__.py
@@ -2,5 +2,5 @@ from .process import (
     Process,
     HandlerFunction
 )
-from . import hyp3_daemon
-from .hyp3_daemon import log
+from . import daemon
+from .daemon import log
