+++ modules/hyp3_process/test/test_hyp3_daemon.py
@@ -12,13 +12,13 @@ import pytest
 from hyp3_events import EmailEvent
 
 import import_hyp3_process
-from hyp3_process.hyp3_daemon import HyP3Daemon, HyP3DaemonConfig, log
-from hyp3_process.hyp3_daemon.worker import WorkerStatus
-from hyp3_process.hyp3_daemon.services import BadMessageException, SQSJob, SQSService
+from hyp3_process.daemon import HyP3Daemon, HyP3DaemonConfig, log
+from hyp3_process.daemon.worker import WorkerStatus
+from hyp3_process.daemon.services import BadMessageException, SQSJob, SQSService
 
 
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.SQSService')
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.HyP3Daemon._process_job')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.SQSService')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.HyP3Daemon._process_job')
 def test_daemon_main(process_job_mock, SQSServiceMock, config, handler):
     log.setLevel(logging.DEBUG)
     daemon = HyP3Daemon(config, handler)
@@ -35,8 +35,8 @@ def test_daemon_main(process_job_mock, SQSServiceMock, config, handler):
     )
 
 
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.SNSService')
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.SQSService')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.SNSService')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.SQSService')
 def test_daemon_main_job_finished(_1, sns_mock, config, handler):
     log.setLevel(logging.DEBUG)
     daemon = HyP3Daemon(config, handler)
@@ -63,10 +63,10 @@ def test_daemon_main_job_finished(_1, sns_mock, config, handler):
 
 
 @pytest.mark.timeout(5)
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.EmailEvent')
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.HyP3Daemon._terminate')
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.SNSService')
-@mock.patch('hyp3_process.hyp3_daemon.hyp3_daemon.SQSService')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.EmailEvent')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.HyP3Daemon._terminate')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.SNSService')
+@mock.patch('hyp3_process.daemon.hyp3_daemon.SQSService')
 def test_shutdown_if_idle(_1, _2, terminate_mock, event_mock, config, handler):
     with pytest.raises(SystemExit):
         daemon = HyP3Daemon(config, handler)
@@ -94,8 +94,8 @@ class MockMessage(object):
 
 
 @mock.patch('boto3.resource')
-@mock.patch('hyp3_process.hyp3_daemon.services.SQSService.validate_message')
-@mock.patch('hyp3_process.hyp3_daemon.services.SQSJob')
+@mock.patch('hyp3_process.daemon.services.SQSService.validate_message')
+@mock.patch('hyp3_process.daemon.services.SQSJob')
 def test_sqsservice_get_next_message(_, validate_mock, sqs_mock):
     sqs_service = SQSService('')
 
@@ -110,7 +110,7 @@ def test_sqsservice_get_next_message(_, validate_mock, sqs_mock):
 
 
 @mock.patch('boto3.resource')
-@mock.patch('hyp3_process.hyp3_daemon.services.SQSJob')
+@mock.patch('hyp3_process.daemon.services.SQSJob')
 def test_sqsservice_get_next_message_bad_checksum(_, sqs_mock):
     sqs_service = SQSService('')
